fix: update controller fot entity
