+++ src/main/java/com/BI/Utils/Status.java
@@ -1,0 +1,8 @@
+package com.BI.Utils;
+
+public enum Status {
+    CREATED,
+    IN_PROGRESS,
+    COMPLETED,
+    SUSPENDED
+}
