+++ src/main/java/com/BI/Utils/TypeTransaction.java
@@ -1,0 +1,10 @@
+package com.BI.Utils;
+
+
+import jakarta.persistence.Table;
+
+
+public enum TypeTransaction {
+    INCOME,
+    EXPENSES
+}
